requirements.txt
django == 5.1.7
djangorestframework == 3.15.2
channels == 4.2.0
psycopg2 == 2.9.10
Pillow == 11.1.0
python-decouple == 3.8
djangorestframework_simplejwt == 5.5.0
redis == 5.2.1
drf-nested-routers == 0.94.1
config == 0.5.1
psycopg2-binary == 2.9.10